+++ src/main/kotlin/net/aquadc/mike/plugin/bigdecimal/BigDecimalConstantInspection.kt
@@ -2,7 +2,9 @@ package net.aquadc.mike.plugin.bigdecimal
 
 import com.intellij.codeInspection.CleanupLocalInspectionTool
 import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.pom.java.LanguageLevel
 import com.intellij.psi.*
+import com.intellij.psi.util.PsiUtil
 import com.siyeh.ig.callMatcher.CallMatcher
 
 import net.aquadc.mike.plugin.FunctionCallVisitor
@@ -16,6 +18,8 @@ import org.jetbrains.uast.UastCallKind
 import org.jetbrains.uast.UastCallKind.Companion.CONSTRUCTOR_CALL
 import org.jetbrains.uast.UastCallKind.Companion.METHOD_CALL
 import org.jetbrains.uast.visitor.AbstractUastNonRecursiveVisitor
+import java.math.BigDecimal
+import java.math.BigInteger
 
 /**
  * BigDecimal instantiation can be replaced with constant
@@ -44,27 +48,33 @@ class BigDecimalConstantInspection : UastInspection(), CleanupLocalInspectionToo
         }
 
         /**
-         * Use constant instead of BigDecimal.valueOf()
+         * Use constant instead of Big(Decimal|Integer).valueOf(long | double)
          */
         private fun visitMethodCall(src: PsiElement, declaringClassFqn: String, args: List<UExpression>) {
             if (CONSTRUCTOR_METHOD.test(src)) {
-                (args.singleOrNull()?.let(UExpression::evaluate) as? Number)
-                    ?.let(::constantOfNumber)?.let { replacement ->
-                        complain(src, declaringClassFqn, replacement)
-                    }
+                (args.singleOrNull()?.let(UExpression::evaluate) as? Number)?.let {
+                    constantOfNumber(it, PsiUtil.getLanguageLevel(src), BI_CONSTRUCTOR_METHOD.test(src))
+                }?.let { replacement ->
+                    complain(src, declaringClassFqn, replacement)
+                }
             }
         }
 
         /**
-         * Use constant instead of new BigDecimal()
+         * Use constant instead of new Big(Decimal|Integer)(String)
          */
         fun visitNewExpression(src: PsiElement, declaringClassFqn: String, args: List<UExpression>) {
             val argVal = args.takeIf { declaringClassFqn.isBigNumber() }?.singleOrNull()?.evaluate() ?: return
 
             when (argVal) {
-                is Number -> constantOfNumber(argVal)
-                is String -> constantOfString(argVal)
-                else -> null // that's ok e.g. 12.34
+                is Number ->
+                    constantOfNumber(argVal, PsiUtil.getLanguageLevel(src), declaringClassFqn == TBigInteger)
+                is String ->
+                    try { constantOfNumber(BigDecimal(argVal).toBigIntegerExact(), PsiUtil.getLanguageLevel(src), declaringClassFqn == TBigInteger) }
+                    catch (e: NumberFormatException) { null }
+                    catch (e: ArithmeticException) { null }
+                else ->
+                    null // that's ok e.g. 12.34
             }?.let { replacement ->
                 complain(src, declaringClassFqn, replacement)
             }
@@ -75,34 +85,19 @@ class BigDecimalConstantInspection : UastInspection(), CleanupLocalInspectionToo
             val unqualified = prefix.substring(prefix.lastIndexOf('.') + 1) + '.' + replacement
             holder.registerProblem(
                 call,
-                "${call.text} should be replaced with $unqualified constant",
-                NamedReplacementFix("Replace instantiation with constant", "$prefix.$replacement")
+                "Instantiation should be replaced with a constant",
+                NamedReplacementFix("Replace ${call.text} with '$unqualified'", "$prefix.$replacement")
             )
         }
 
-        private fun constantOfNumber(number: Number): String? {
-            if (!number.toDouble().isInt) {
-                // that's ok
-                return null
-            }
-            return when (number.toInt()) {
-                0 -> "ZERO"
-                1 -> "ONE"
-                // TODO "TWO" for BigInteger and Java 9
-                10 -> "TEN"
+        private fun constantOfNumber(number: Number, languageLevel: LanguageLevel, isBigInt: Boolean): String? {
+            return when (number) {
+                0, 0L, 0.0, BigInteger.ZERO -> "ZERO"
+                1, 1L, 1.0, BigInteger.ONE -> "ONE"
+                2, 2L, 2.0, BigInteger.TWO -> "TWO".takeIf { isBigInt && languageLevel >= LanguageLevel.JDK_1_9 }
+                10, 10L, 10.0, BigInteger.TEN -> "TEN"
                 else -> null // that's ok
             }
-        }
-
-        val Double.isInt: Boolean
-            get() = this == toInt().toDouble()
-
-        private fun constantOfString(str: String): String? = when (str) {
-            "0" -> "ZERO"
-            "1" -> "ONE"
-            // TODO "TWO" for BigInteger and Java 9
-            "10" -> "TEN"
-            else -> null // that's ok
         }
     }
 
@@ -113,10 +108,8 @@ fun String.isBigNumber() = this == TBigDecimal || this == TBigInteger
 const val TBigDecimal = "java.math.BigDecimal"
 const val TBigInteger = "java.math.BigInteger"
 
-private val CONSTRUCTOR_METHOD = CallMatcher.staticCall(TBigDecimal, "valueOf").let {
-    CallMatcher.anyOf(
-        it.parameterTypes("long"),
-        it.parameterTypes("double"),
-        CallMatcher.staticCall(TBigInteger, "valueOf").parameterTypes("long"),
-    )
+private val BD_CONSTRUCTOR_METHOD = CallMatcher.staticCall(TBigDecimal, "valueOf").let {
+    CallMatcher.anyOf(it.parameterTypes("long"), it.parameterTypes("double"))
 }
+private val BI_CONSTRUCTOR_METHOD = CallMatcher.staticCall(TBigInteger, "valueOf").parameterTypes("long")
+private val CONSTRUCTOR_METHOD = CallMatcher.anyOf(BD_CONSTRUCTOR_METHOD, BI_CONSTRUCTOR_METHOD)