+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   <ul>
     <li>Added live templates for implementing <code>Property</code> for <code>ObjectAnimator</code></li>
     <li>Reporting that <code>Activity#onCreate(, PersistableBundle)</code> will highly likely not be called</li>
+    <li>Reporting that <code>&lt;AnyScrollableView></code> should have an ID to save its scroll position</li>
   </ul>
 
 ### 0.27